Bug fix for stabilizing the regression suite in Distributor Part seven
