+++ src/test/java/com/cutanddry/qa/tests/settings/ValidateTheFlowOfTheAddNewWarehouseUnderTheManageWarehouseOptionTest.java
@@ -12,9 +12,8 @@ import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 import org.testng.asserts.SoftAssert;
 
-public class VerifyTheFlowOfEditingTheConfigurationSettingsForTheNewlyAddedWarehouse extends TestBase{
+public class ValidateTheFlowOfTheAddNewWarehouseUnderTheManageWarehouseOptionTest extends TestBase {
     static User user;
-    static String nickname = "Test Nickname";
 
     @BeforeMethod
     public void setUp() {
@@ -22,8 +21,8 @@ public class VerifyTheFlowOfEditingTheConfigurationSettingsForTheNewlyAddedWareh
         user = JsonUtil.readUserLogin();
     }
 
-    @Test(groups = "DOT-TC-856")
-    public void VerifyTheFlowOfAddingTheConfigurationSettingsForTheNewlyAddedWarehouse() throws InterruptedException {
+    @Test(groups = "DOT-TC-852")
+    public void ValidateTheFlowOfTheAddNewWarehouseUnderTheManageWarehouseOption() throws InterruptedException {
         SoftAssert softAssert = new SoftAssert();
         Login.loginAsDistributor(user.getEmailOrMobile(), user.getPassword());
         Dashboard.isUserNavigatedToDashboard();
@@ -39,28 +38,12 @@ public class VerifyTheFlowOfEditingTheConfigurationSettingsForTheNewlyAddedWareh
         Settings.selectWarehouseLocation(warehouseName);
         String lastFiveDigits = warehouseName.substring(warehouseName.length() - 5);
         softAssert.assertTrue(Settings.isWarehouseNameDisplayed(lastFiveDigits),"The warehouse name is not displayed");
-
-        Login.switchIntoNewTab();
-        Login.navigateToDistributor();
-        Login.loginAsDistributor(user.getEmailOrMobile(), user.getPassword());
-        softAssert.assertTrue(Dashboard.isUserNavigatedToDashboard(),"login error");
-        Dashboard.navigateToTrackSettings();
-        softAssert.assertTrue(Settings.isTrackSettingsTextDisplayed(),"navigation error");
-        Settings.clickOnWarehouseLocationDropDown();
-        Settings.selectWarehouseLocation(warehouseName);
-        softAssert.assertTrue(Settings.isWarehouseNameDisplayed(lastFiveDigits),"The warehouse name is not displayed");
-        Settings.fillWarehouseName(warehouseName);
-        Settings.fillWarehouseNickName(nickname);
-        Settings.clickOnSaveChanges();
-//        softAssert.assertTrue(Settings.isUpdatedSuccessfullyDisplayed(),"Message not displayed");
         softAssert.assertAll();
     }
-
 
     @AfterMethod
     public void tearDown(ITestResult result) {
         takeScreenshotOnFailure(result);
         closeAllBrowsers();
     }
-
 }
