+++ src/test/java/com/cutanddry/qa/tests/settings/VerifyTheFlowOfAddingTheConfigurationSettingsForTheNewlyAddedWarehouseTest.java
@@ -12,8 +12,16 @@ import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 import org.testng.asserts.SoftAssert;
 
-public class ValidateTheFlowOfTheAddNewWarehouseUnderTheManageWarehouseOption extends TestBase {
+public class VerifyTheFlowOfAddingTheConfigurationSettingsForTheNewlyAddedWarehouseTest extends TestBase {
     static User user;
+    static String nickname = "Test Nickname";
+    static String streetAddress = "123 Test Street";
+    static String apartment = "Apt 1";
+    static String city = "Test City";
+    static String state = "Test State";
+    static String zip = "12345";
+    static String alertEmailInput = "test@email.com";
+    static String supportEmailInput = "test@email.com";
 
     @BeforeMethod
     public void setUp() {
@@ -21,8 +29,8 @@ public class ValidateTheFlowOfTheAddNewWarehouseUnderTheManageWarehouseOption ex
         user = JsonUtil.readUserLogin();
     }
 
-    @Test(groups = "DOT-TC-852")
-    public void ValidateTheFlowOfTheAddNewWarehouseUnderTheManageWarehouseOption() throws InterruptedException {
+    @Test(groups = "DOT-TC-854")
+    public void VerifyTheFlowOfAddingTheConfigurationSettingsForTheNewlyAddedWarehouse() throws InterruptedException {
         SoftAssert softAssert = new SoftAssert();
         Login.loginAsDistributor(user.getEmailOrMobile(), user.getPassword());
         Dashboard.isUserNavigatedToDashboard();
@@ -38,8 +46,22 @@ public class ValidateTheFlowOfTheAddNewWarehouseUnderTheManageWarehouseOption ex
         Settings.selectWarehouseLocation(warehouseName);
         String lastFiveDigits = warehouseName.substring(warehouseName.length() - 5);
         softAssert.assertTrue(Settings.isWarehouseNameDisplayed(lastFiveDigits),"The warehouse name is not displayed");
+
+        Login.switchIntoNewTab();
+        Login.navigateToDistributor();
+        Login.loginAsDistributor(user.getEmailOrMobile(), user.getPassword());
+        softAssert.assertTrue(Dashboard.isUserNavigatedToDashboard(),"login error");
+        Dashboard.navigateToTrackSettings();
+        softAssert.assertTrue(Settings.isTrackSettingsTextDisplayed(),"navigation error");
+        Settings.clickOnWarehouseLocationDropDown();
+        Settings.selectWarehouseLocation(warehouseName);
+        softAssert.assertTrue(Settings.isWarehouseNameDisplayed(lastFiveDigits),"The warehouse name is not displayed");
+        Settings.fillWarehouseInfo(nickname, warehouseName,  streetAddress,  apartment,  city,  state,  zip,  alertEmailInput,  supportEmailInput);
+        Settings.clickOnSaveChanges();
+//        softAssert.assertTrue(Settings.isUpdatedSuccessfullyDisplayed(),"Message not displayed");
         softAssert.assertAll();
     }
+
 
     @AfterMethod
     public void tearDown(ITestResult result) {
